feat: delegate layout
