+++ src/app/delegate/layout.tsx
@@ -1,0 +1,25 @@
+import Header from '@/components/header'
+import { client } from '@/gql/client'
+import { siteDataQuery } from '@/gql/query'
+
+import { DelegateFooter } from '@/components/delegate/delegate-footer'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Delegate',
+}
+export const fetchCache = 'force-cache'
+
+const DelegateLayout = async ({ children }: { children: React.ReactNode }) => {
+  const cmsData = await client.request(siteDataQuery)
+
+  return (
+    <>
+      <Header cms={cmsData} />
+      {children}
+      <DelegateFooter />
+    </>
+  )
+}
+
+export default DelegateLayout
